Camera Follow, Map Update, Continue Blackout
Camera now follows the player if they are on a tagged "map".
The maps get updated for the current scene with the MapSaver script
(adds a 2D collider to the size of the object) if tagged with "map",
assumed from Tiled2Unity
When continuing game, waits a second before fading back in so you don't
see the player warp.